refactor sv
