+++ philo/srcs/actors/sv/_on_actor_start_done.c
@@ -1,0 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   _on_actor_start_done.c                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/17 21:50:17 by ttsubo            #+#    #+#             */
+/*   Updated: 2025/06/21 17:31:27 by ttsubo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "sv_actor.h"
+
+void	_on_actor_start_done(t_sv_actor *self)
+{
+	self->prop->start_done_count++;
+	if (self->prop->start_done_count == self->prop->ref_count)
+	{
+		_send_grant_eat(self);
+		gettimeofday(&self->prop->start, NULL);
+		self->prop->last_update_time = self->prop->start;
+		self->sys_notify_inbox->enq(
+			self->sys_notify_inbox, msg_new(ACTOR_START_DONE, NULL, NULL));
+	}
+}
