+++ philo/srcs/actors/sv/_on_philo_dead.c
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   _on_philo_dead.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/17 21:50:17 by ttsubo            #+#    #+#             */
+/*   Updated: 2025/06/21 17:24:17 by ttsubo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "sv_actor.h"
+
+void	_on_philo_dead(t_sv_actor *self)
+{
+	t_ft_queue	*notify;
+
+	notify = self->sys_notify_inbox;
+	notify->enq(notify, msg_new(SYSTEM_STOP, NULL, NULL));
+}
