+++ philo/srcs/actors/sv/sv_actor.c
@@ -6,29 +6,57 @@
 /*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/03 18:42:16 by ttsubo            #+#    #+#             */
-/*   Updated: 2025/06/21 10:00:40 by ttsubo           ###   ########.fr       */
+/*   Updated: 2025/06/21 17:30:38 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "main.h"
 #include "sv_actor.h"
 
-static void	sv_on_start(t_ft_actor *self)
+static void	_sv_prop_del(t_sv_prop **prop_ptr)
 {
-	size_t		i;
+	t_sv_prop	*prop;
+
+	if (!prop_ptr || !*prop_ptr)
+		return ;
+	prop = *prop_ptr;
+	free(prop->philos_ref);
+	free(prop->forks_ref);
+	free_matrix(&prop->send_ptn);
+	free(prop);
+	*prop_ptr = NULL;
+}
+
+static t_sv_prop	*_sv_prop_new(t_main_args args)
+{
+	t_sv_prop	*prop;
+
+	prop = philo_calloc(1, sizeof(t_sv_prop));
+	if (!prop)
+		return (NULL);
+	prop->args = args;
+	prop->philos_ref = philo_calloc(args.num_of_philos, sizeof(t_ft_actor *));
+	prop->forks_ref = philo_calloc(args.num_of_philos, sizeof(t_ft_actor *));
+	prop->philo_done_count = 0;
+	prop->ptn_i = 0;
+	prop->send_ptn = _create_send_ptn(args.num_of_philos);
+	if (!prop->send_ptn || !prop->philos_ref)
+		_sv_prop_del(&prop);
+	return (prop);
+}
+
+void	sv_actor_del(t_sv_actor **sv_ptr)
+{
 	t_sv_actor	*sv;
 
-	sv = self->ref;
-	i = 0;
-	while (i < sv->prop->args.num_of_philos)
-	{
-		ft_actor_start(sv->prop->philos_ref[i]);
-		ft_actor_start(sv->prop->forks_ref[i]);
-		sv->prop->ref_count += 2;
-		i++;
-	}
-	ft_actor_start(sv->prop->monitor_ref);
-	sv->prop->ref_count++;
+	sv = *sv_ptr;
+	if (!sv)
+		return ;
+	ft_actor_del(&sv->base);
+	free_matrix(&sv->prop->send_ptn);
+	_sv_prop_del(&sv->prop);
+	free(sv);
+	*sv_ptr = NULL;
 }
 
 t_sv_actor	*sv_actor_new(t_main_args args)
@@ -36,7 +64,7 @@ t_sv_actor	*sv_actor_new(t_main_args args)
 	t_sv_actor						*sv;
 	static const t_ft_actor_vtable	vtable = {
 		.on_start = sv_on_start,
-		.on_receive = _sv_on_receive,
+		.on_receive = sv_on_receive,
 		.on_stop = NULL,
 	};
 
@@ -46,5 +74,7 @@ t_sv_actor	*sv_actor_new(t_main_args args)
 	sv->base = ft_actor_new(sv);
 	sv->base->v = &vtable;
 	sv->prop = _sv_prop_new(args);
+	if (!sv->base || !sv->prop)
+		sv_actor_del(&sv);
 	return (sv);
 }
